Don’t write to .bundle/config (#49)
Instead or writing to .bundle/config we can instead us the environment variable interface present to define bundler config vars. I wrote about this more here: https://devcenter.heroku.com/articles/bundler-configuration#environment-variable-behavior

This should provide the same behavior as before. 

I did notice one issue with the current implementation and that is the values in .bundle/config that are written are the tmp dir paths from build time and not the paths to /app for runtime, though I don’t think it matters, it can’t hurt to fix it.